+++ src/vpux_translate_utils/src/hwtest/hwtest_utils.cpp
@@ -318,20 +318,20 @@ mlir::MemRefType getMemRefType(VPURT::BufferSection section, ArrayRef<int64_t> s
 
 vpux::VPURT::DeclareBufferOp createDeclareTensorOp(mlir::OpBuilder& builder, VPURT::BufferSection section,
                                                    ArrayRef<int64_t> shape, mlir::Type elemType, DimsOrder order,
-                                                   int locale, size_t offset) {
+                                                   int64_t locale, size_t offset) {
     const auto type = getMemRefType(section, shape, elemType, order);
     return builder.create<VPURT::DeclareBufferOp>(builder.getUnknownLoc(), type, section, locale, offset);
 }
 
 vpux::VPURT::DeclareBufferOp createDeclareTensorOp(mlir::OpBuilder builder, VPURT::BufferSection section,
                                                    ArrayRef<int64_t> shape, mlir::Type elemType, DimsOrder order,
-                                                   StridesRef strides, int locale, size_t offset) {
+                                                   StridesRef strides, int64_t locale, size_t offset) {
     const auto type = getMemRefType(section, shape, elemType, order, strides);
     return builder.create<VPURT::DeclareBufferOp>(builder.getUnknownLoc(), type, section, locale, offset);
 }
 
 vpux::VPURT::DeclareBufferOp createDeclareTensorOp(mlir::OpBuilder& builder, mlir::MemRefType type,
-                                                   VPURT::BufferSection section, int locale, size_t offset) {
+                                                   VPURT::BufferSection section, int64_t locale, size_t offset) {
     return builder.create<VPURT::DeclareBufferOp>(builder.getUnknownLoc(), type, section, locale, offset);
 }
